CZN-29: Update domain model class.
