+++ src/api/mapper/user_mapper.py
@@ -1,7 +1,6 @@
-from src.api.model.address import Address
+from api.model.domain import Address, User
 from src.api.model.schemas import Address as UserRegistrationRequestAddress
 from src.api.model.schemas import UserRegistrationRequest, UserResponse
-from src.api.model.user import User
 
 
 class UserMapper:
