+++ src/api/model/domain.py
@@ -1,8 +1,25 @@
 from datetime import datetime, timezone
 from typing import Optional
 
-from src.api.model.address import Address
 from src.api.model.enum import UserRole, UserStatus
+
+
+class Address:
+    def __init__(
+        self,
+        id: Optional[int] = None,
+        street: Optional[str] = None,
+        city: Optional[str] = None,
+        state: Optional[str] = None,
+        country: Optional[str] = None,
+        postal_code: Optional[str] = None,
+    ):
+        self.id = id
+        self.street = street
+        self.city = city
+        self.state = state
+        self.country = country
+        self.postal_code = postal_code
 
 
 class User:
