+++ src/api/repository/user_repository.py
@@ -4,10 +4,9 @@ from fastapi import HTTPException, status
 from psycopg2 import errors
 from psycopg2.extras import DictCursor
 
+from api.model.domain import Address, User
 from src.api.config.database import DatabasePool
 from src.api.mapper.user_mapper import UserMapper
-from src.api.model.address import Address
-from src.api.model.user import User
 
 
 class UserRepository:
